pyproject.toml

[tool.poetry]
name = "eink"
version = "0.1.0"
description = ""
authors = []
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
gpiozero = "^1.5.1"
spidev = "^3.5"
Pillow = "^8.0.1"
requests = "^2.25.1"
PyYAML = "^5.3.1"
todoist-python = "^8.1.3"
python-dateutil = "^2.8.1"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
